pyproject.toml
[project]
name = "minutes-creation-tool"
version = "0.1.0"
description = "Tool for creating AI meeting minutes using transcription and summarization"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "assemblyai",
    "langchain",
    "langchain-openai",
    "langchain-anthropic",
    "anthropic",
    "python-dotenv",
    "pydantic",
    "transformers>=4.49.0",
]
